move glob pattern matching into teamscale-client to make it reusable
